refactor: move to __tests__ folder
